+++ src/string_utils.h
@@ -1,3 +1,7 @@
+
+#ifndef STRINGUTILS_GUARD
+#define STRINGUTILS_GUARD
+
 #include<string>
 #include<vector>
 
@@ -32,4 +36,7 @@ void make_valid_names(std::vector<std::string> &s);
 
 void str_replace(std::string& str, const std::string& from, const std::string& to);
 size_t str_replace_all(std::string& str, const std::string& from, const std::string& to);
+
+
+#endif